+++ arch/src/element.test.ts
@@ -1,7 +1,9 @@
-const {suite, test} = intern.getInterface('tdd')
-const {assert} = intern.getPlugin('chai')
+import 'intern'
 import {element} from './element'
 import {h} from '../../dom/src/html/h'
+
+const {suite, test} = intern.getInterface('tdd')
+const {assert} = intern.getPlugin('chai')
 
 suite('element()', () => {