+++ arch/src/extend.ts
@@ -5,7 +5,11 @@
 export function extend<T, U>(target: T, source: U): T & U {
     for (const prop in source) {
         if (source.hasOwnProperty(prop)) {
-            Object.defineProperty(target, prop, Object.getOwnPropertyDescriptor(source, prop)!)
+            Object.defineProperty(
+                target,
+                prop,
+                Object.getOwnPropertyDescriptor(source, prop)!
+            )
         }
     }
     return target as T & U