+++ arch/src/index.test.ts
@@ -1,12 +1,14 @@
-const {suite, test} = intern.getInterface('tdd')
+import 'intern'
+import {assert, chan, element, extend} from './index'
 
-import {assert, chan, extend, Component as _C, ComponentFactory as _CF} from './index'
+const {suite, test} = intern.getInterface('tdd')
 
 suite('index (@soil/arch)', () => {
 
     test('exports the right artifacts', () => {
         assert(true, '')
         chan<any>()
+        element(() => [document.createElement('div'), {}])
         extend({}, {})
     })
 })