+++ arch/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 1.0.0
 
 ### Added
-- `Component` and `ComponentFactory` types, encoding the shape of custom components.
-- `chan()` function, helping with communications between distant components.
 - `assert()` function, for basic contract programming support.
-- `extend()` function, to facilitate the extension of HTML elements with custom properties, including getters and setters.
+- `chan()` function, helping with communications between distant components.
+- `element()` function, to define custom components similar to native ones.
+- `extend()` function, to facilitate the extension of objects with custom
+  properties, including getters and setters.